% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgage.R
\name{avgage}
\alias{avgage}
\title{Average Age Function}
\usage{
avgage(x)
}
\arguments{
\item{x}{The year you are inquiring.}
}
\description{
This function allows you to inquire the average age of NBA players in any certain year.
}
\examples{
avgage()
}
\keyword{age}
