% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{Correlation Function}
\usage{
correlation(x)
}
\arguments{
\item{x}{The year you are inquiring.}
}
\description{
This function gives a correlation panel of all numerical data in a certain year, showing the correlations between.
}
\examples{
correlation()
}
\keyword{correlation}
