% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest.R
\name{oldest}
\alias{oldest}
\title{Oldest Player Function}
\usage{
oldest(x)
}
\arguments{
\item{x}{The year you are inquiring.}
}
\description{
This function allows you to inquire the age of the oldest NBA players in a certain year.
}
\examples{
oldest()
}
\keyword{oldest}
